Simpler Highlight cmd and switch to custom gruvbox
The statusline colour manipulations will only apply if there are no
changes to the `User1` group, this is so that I can set my own group
settings in the color script but still have a fallback for other color
scripts.

Increased number of files recorded in viminfo (do I need to set
shadainfo too for nvim?).

The Highlight cmd takes a single arg that pattern matches on the whole
group description.

Improvements to DiffOrig to set correct ft on original buffer and maps
q to quickly exit diff. Also <leader>ud for DiffOrig.